refactor: KitchenridersClient 패키지 변경
