+++ src/main/java/kitchenpos/order/application/KitchenridersClient.java
@@ -1,4 +1,4 @@
-package kitchenpos.kitchenriders;
+package kitchenpos.order.application;
 
 import java.math.BigDecimal;
 import java.util.UUID;
