+++ src/main/java/kitchenpos/order/infra/DefaultKitchenridersClient.java
@@ -1,5 +1,6 @@
-package kitchenpos.kitchenriders;
+package kitchenpos.order.infra;
 
+import kitchenpos.order.application.KitchenridersClient;
 import org.springframework.stereotype.Component;
 
 import java.math.BigDecimal;
